requirements.txt
binance-connector==3.7.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
